# DB/OpenAPIList.yaml
openapi: 3.0.0
info:
  title: API di Gestione Utenti e Richieste
  description: API per la gestione del progetto WebShop di Sviluppo Web Avanzato.
  version: 1.0.0
servers:
  - url: http://localhost:8080/disim_swa-1.0/
    description: Server locale
paths:
...
  /rest/categorie:
    get:
      summary: Recupera lista categorie disponibili
      responses:
        '200':
          description: Lista categorie
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '404':
          description: Nessuna categoria trovata

    put:
      summary: Modifica categoria
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Categoria'
      responses:
        '204':
          description: Categoria aggiornata con successo
        '404':
          description: Categoria non trovata
        '400':
          description: Non sei l'amministratore
        '500':
          description: Errore generico
    
    delete:
      summary: Elimina categoria
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '204':
          description: Categoria eliminata con successo
        '400':
          description: Non sei l'amministratore
        '404':
          description: Categoria non trovata
        '500':
          description: Errore generico 
  /rest/categorie/genitori:
    get:
      summary: Recupera tutte le categorie genitori (senza genitore_padre)
      responses:
        '200':
          description: Lista categorie genitori
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '404':
          description: Nessuna categoria trovata
  /rest/categorie/piuvendute:
    get:
      summary: Recupera le 3 categorie più vendute
      responses:
        '200':
          description: Lista categorie più vendute
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '404':
          description: Nessuna categoria trovata
...
  /rest/immagini:
    get:
      summary: Recupera lista immagini
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista immagini
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Immagine' 
        '404':
          description: Nessuna immagine trovata
  
    put:
      summary: Carica l'immagine
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                immagine:
                  type: string
                  format: binary
                titolo:
                  type: string
      responses:
        '204':
          description: Immagine caricata con successo
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Non sei l'amministratore
        '404':
          description: Immagine non trovata
        '409':
          description: Caricamento immagine fallito
        '500':
          description: Errore generico
          
    delete:
      summary: Cancella immagine
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '204': 
          description: Immagine cancellata con successo
        '400':
          description: Non sei l'amministratore
        '404':
          description: Immagine non trovata
        '500':
          description: Errore generico
...
  /rest/immagini/categoria:
    get:
      summary: Recupera l'immagine in base alla categoria
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Dettagli immagine 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Immagine' 
        '404':
          description: Nessuna immagine trovata     
  /rest/immagini/download:
    get:
      summary: Restituisce l'immagine da visualizzare
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Binario dell'immagine 
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                description: Used only with `image/jpeg` and `image/png` responses
                example: attachment; filename="name.pdf"
        '404':
          description: Immagine non trovata  
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Categoria:
      type: object
      properties:
        id: 
          type: integer
          example: 1
        nome:
          type: string
          example: "Pc"
        immagine: 
          type: object
          items:
            $ref: "#/components/schemas/Immagine"
        categoria_padre:
          type: object
          items:
            $ref: "#/components/schemas/Categoria"
        caratteristiche:
          type: array
          items:
            $ref: "#/components/schemas/Caratteristica"
        cancellato:
          type: boolean
          example: false
    Caratteristica:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: "Memoria RAM"
        categoria:
          type: object
          items:
            $ref: "#/components/schemas/Categoria"
        valori_default:
          type: string
          example: "12Gb, 24Gb, 36Gb"
    Immagine:
      type: object
      properties:
        id:
          type: integer
          example: 1
        didascalia:
          type: string
          example: "Immagine pc"
        tipo_immagine:
          type: string
          example: "jpeg"
        nome_file_immagine:
          type: string
        grandezza_file:
          type: number
          example: 2000000
...